backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# PDF parsing
pymupdf==1.24.13

# NLP / skill extraction
scikit-learn==1.5.2
numpy==1.26.4

# AI suggestions
openai==1.54.4

# Utilities
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2

# BCrypt (pinned for passlib compat)
bcrypt==4.0.1

# Email validation
email-validator==2.2.0

# Async concurrency
greenlet==3.1.1
